Separate lsps0 schema's
The `crate::lsps::message`-module was becoming too big.
I've decided to separate all files related to LSPS0.

The code should better reflect how the LSP-specification is structured.

I made the following changes
- renamed DateTime to IsoDatetime to prevent some compiler bugs
- clippy fixes

Follow-up commits for LSPS1 and LSPS2 will follow
