+++ libs/gl-client/src/lsps/lsps0/mod.rs
@@ -1,0 +1,2 @@
+pub mod common_schemas;
+pub mod schema;
