+++ libs/gl-client/src/lsps/mod.rs
@@ -4,3 +4,4 @@ pub mod json_rpc;
 pub mod json_rpc_erased;
 pub mod message;
 pub mod client;
+pub mod lsps0;
